% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estaciones_dmc}
\alias{estaciones_dmc}
\title{Estaciones agromet}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 112 rows and 11 columns.
}
\usage{
estaciones_dmc
}
\description{
Estaciones agromet
}
\section{Variables}{


\itemize{

\item \code{ codigoNacional } : codigoNacional.
\item \code{ codigoOMM } : codigoOMM.
\item \code{ codigoOACI } : codigoOACI.
\item \code{ nombreEstacion } : nombreEstacion.
\item \code{ latitud } : latitud.
\item \code{ longitud } : longitud.
\item \code{ altura } : altura.
\item \code{ region } : region.
\item \code{ NombreRegion } : NombreRegion.
\item \code{ zonaGeografica } : zonaGeografica.
\item \code{ datos } : datos.

}
}

\keyword{datasets}
