% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agro_data.R
\name{get_agro_data}
\alias{get_agro_data}
\alias{get_agro_data_from_api}
\title{Get data from agromet.cl API}
\usage{
get_agro_data(
  stations_id = NULL,
  date_start = NULL,
  date_end = NULL,
  verbose = FALSE
)

get_agro_data_from_api(station_id = NULL, date_start = NULL, date_end = NULL)
}
\arguments{
\item{stations_id}{A numeric value indicating stations ID from \code{estaciones_agromet}.}

\item{date_start}{A value parseable by \code{lubridate::as_datetime}, for example \code{"2020-11-30 01:00:00"}.}

\item{date_end}{Same as \code{date_start} parameter.}

\item{verbose}{A logical value to show or not a progress bar with eta.}

\item{station_id}{A numeric length 1 value indicating station ID from \code{estaciones_agromet}.}
}
\description{
Get data from agromet.cl API
}
\examples{

get_agro_data(1:10, "2020-11-30 01:00:00", "2020-12-15 01:00:00")

get_agro_data(1:10, "2020-11-30 01:00:00", "2020-12-15 01:00:00", verbose = TRUE)


get_agro_data_from_api(108, "2020-11-30 01:00:00", "2020-12-15 01:00:00")

}
