% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agro_data.R
\name{get_agro_data_from_api}
\alias{get_agro_data_from_api}
\title{Get data from agromet.cl API}
\usage{
get_agro_data_from_api(station_id = NULL, date_start = NULL, date_end = NULL)
}
\description{
Get data from agromet.cl API
}
\examples{

get_agro_data_from_api(108, "2020-11-30 01:00:00", "2020-12-15 01:00:00")

}
