% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables_agromet}
\alias{variables_agromet}
\title{Variables_agromet}
\format{
A \code{data frame} with 12 observations and 3 variables.
}
\usage{
variables_agromet
}
\description{
Variables_agromet
}
\section{Variables}{


\itemize{

\item \code{columna}: Nombre de la columna.
\item \code{unit}: Unidad de medida.
\item \code{description}: Descripción de la columna.

}
}

\keyword{datasets}
