% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estaciones_agromet}
\alias{estaciones_agromet}
\title{Estaciones agromet}
\format{
A \code{data frame} with 417  observations and 8 variables.
}
\usage{
estaciones_agromet
}
\description{
Estaciones agromet
}
\section{Variables}{


\itemize{#'
\item \code{ ema } : ema.
\item \code{ institucion } : institucion.
\item \code{ nombre_ema } : nombre_ema.
\item \code{ comuna } : comuna.
\item \code{ region } : region.
\item \code{ latitud } : latitud.
\item \code{ longitud } : longitud.
\item \code{ fecha_de_alta } : fecha_de_alta.
}
}

\keyword{datasets}
